+++ crates/ruma-client-api/CHANGELOG.md
@@ -11,6 +11,7 @@ Breaking changes:
   * Move `push::get_pushers::v3::Pusher` to `push` and make it use the new `PusherIds` type
   * Remove `push::set_pusher::v3::Pusher` and use the common type instead
 * Make `push::PusherKind` contain the pusher's `data`
+* Use an enum for the `scope` of the `push` endpoints
 
 Improvements: